+++ examples/power.c
@@ -11,24 +11,24 @@
 #include "hal_power.h"
 
 int main() {
-    enum power_sleep_modes sleep_mode;
-    enum power_modules module;
+    enum hal_power_sleep_modes sleep_mode;
+    enum hal_power_modules module;
 
     // Setting general CPU power.
-    sleep_mode = power_idle_mode;
-    power_set_sleep_mode(sleep_mode);
+    sleep_mode = hal_power_idle_mode;
+    hal_power_set_sleep_mode(sleep_mode);
 
     // Disabling ADC.
-    module = power_adc;
-    power_set_module_power(module, 0);
+    module = hal_power_adc;
+    hal_power_set_module_power(module, 0);
 
     // Disabling SPI.
-    module = power_spi;
-    power_set_module_power(module, 0);
+    module = hal_power_spi;
+    hal_power_set_module_power(module, 0);
 
     // Enabling SPI.
-    module = power_spi;
-    power_set_module_power(module, 1);
+    module = hal_power_spi;
+    hal_power_set_module_power(module, 1);
 
     return 0;
 }