+++ CHANGELOG.md
@@ -6,6 +6,20 @@ also be mirrored to the Github releases page.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Callback function to mock sleep instruction.
+
+### Changed
+
+- Naming conventions for module members like functions.
+
+### Fixed
+
+- Reserved and illegal value usage checks in power module.
+
 ## [0.4.0] - 2025-10-15
 
 ### Added